requirements.txt
certifi==2018.11.29
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.7
django-dotenv==1.4.2
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
drf-yasg==1.11.0
flex==6.13.2
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
MarkupSafe==1.1.0
psycopg2-binary==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rfc3987==1.3.8
ruamel.yaml==0.15.88
six==1.12.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
whitenoise==4.1.2
